update
đã đẩy được các bài post lên giao điẹn

+++ Moe/src/main/java/com/moe/music/jpa/ActivityLogJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.ActivityLog;
 
-public interface ActivityLogJPA extends JpaRepository<ActivityLog, Integer>{
+public interface ActivityLogJPA extends JpaRepository<ActivityLog, Long>{
 
 }
