+++ Moe/src/main/java/com/moe/music/jpa/CommentJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Comment;
 
-public interface CommentJPA extends JpaRepository<Comment, Integer> {
+public interface CommentJPA extends JpaRepository<Comment, Long> {
 
 }
