+++ Moe/src/main/java/com/moe/music/jpa/FeedbackJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Feedback;
 
-public interface FeedbackJPA extends JpaRepository<Feedback, Integer>{
+public interface FeedbackJPA extends JpaRepository<Feedback, Long>{
 
 }
