+++ Moe/src/main/java/com/moe/music/jpa/FollowerJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Follower;
 
-public interface FollowerJPA extends JpaRepository<Follower, Integer>{
+public interface FollowerJPA extends JpaRepository<Follower, Long>{
 
 }
