+++ Moe/src/main/java/com/moe/music/jpa/LikeJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Like;
 
-public interface LikeJPA extends JpaRepository<Like, Integer>{
+public interface LikeJPA extends JpaRepository<Like, Long>{
 
 }
