+++ Moe/src/main/java/com/moe/music/jpa/MessageJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Message;
 
-public interface MessageJPA extends JpaRepository<Message, Integer>{
+public interface MessageJPA extends JpaRepository<Message, Long>{
 
 }
