+++ Moe/src/main/java/com/moe/music/jpa/NotificationJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Notification;
 
-public interface NotificationJPA extends JpaRepository<Notification, Integer>{
+public interface NotificationJPA extends JpaRepository<Notification, Long>{
 
 }
