+++ Moe/src/main/java/com/moe/music/jpa/ReportJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Report;
 
-public interface ReportJPA extends JpaRepository<Report, Integer>{
+public interface ReportJPA extends JpaRepository<Report, Long>{
 
 }
