+++ Moe/src/main/java/com/moe/music/jpa/RoleJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Role;
 
-public interface RoleJPA extends JpaRepository<Role, Integer> {
+public interface RoleJPA extends JpaRepository<Role, Long> {
 
 }
