+++ Moe/src/main/java/com/moe/music/jpa/TagJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moe.music.model.Tag;
 
-public interface TagJPA extends JpaRepository<Tag, Integer> {
+public interface TagJPA extends JpaRepository<Tag, Long> {
 
 }
